+++ Sources/Sentry/SentryCrashIntegration.m
@@ -5,6 +5,7 @@
 #import "SentryEvent.h"
 #import "SentryFrameInAppLogic.h"
 #import "SentryHub.h"
+#import "SentryOutOfMemoryLogic.h"
 #import "SentrySDK.h"
 #import "SentryScope+Private.h"
 #import "SentrySessionCrashedHandler.h"
@@ -20,6 +21,7 @@ SentryCrashIntegration ()
 
 @property (nonatomic, weak) SentryOptions *options;
 @property (nonatomic, strong) SentryDispatchQueueWrapper *dispatchQueueWrapper;
+@property (nonatomic, strong) SentryCrashAdapter *crashAdapter;
 @property (nonatomic, strong) SentrySessionCrashedHandler *crashedSessionHandler;
 
 @end
@@ -29,10 +31,8 @@ SentryCrashIntegration ()
 - (instancetype)init
 {
     if (self = [super init]) {
-        SentryCrashAdapter *crashWrapper = [[SentryCrashAdapter alloc] init];
+        self.crashAdapter = [[SentryCrashAdapter alloc] init];
         self.dispatchQueueWrapper = [[SentryDispatchQueueWrapper alloc] init];
-        self.crashedSessionHandler =
-            [[SentrySessionCrashedHandler alloc] initWithCrashWrapper:crashWrapper];
     }
     return self;
 }
@@ -42,9 +42,8 @@ SentryCrashIntegration ()
              andDispatchQueueWrapper:(SentryDispatchQueueWrapper *)dispatchQueueWrapper
 {
     self = [self init];
+    self.crashAdapter = crashAdapter;
     self.dispatchQueueWrapper = dispatchQueueWrapper;
-    self.crashedSessionHandler =
-        [[SentrySessionCrashedHandler alloc] initWithCrashWrapper:crashAdapter];
 
     return self;
 }
@@ -65,6 +64,13 @@ SentryCrashIntegration ()
 - (void)installWithOptions:(nonnull SentryOptions *)options
 {
     self.options = options;
+
+    SentryOutOfMemoryLogic *logic =
+        [[SentryOutOfMemoryLogic alloc] initWithOptions:options crashAdapter:self.crashAdapter];
+    self.crashedSessionHandler =
+        [[SentrySessionCrashedHandler alloc] initWithCrashWrapper:self.crashAdapter
+                                                 outOfMemoryLogic:logic];
+
     [self startCrashHandler];
     [self configureScope];
 }
@@ -93,7 +99,7 @@ SentryCrashIntegration ()
         // there and the AutoSessionTrackingIntegration can work properly.
         //
         // This is a pragmatic and not the most optimal place for this logic.
-        [self.crashedSessionHandler endCurrentSessionAsCrashedWhenCrashed];
+        [self.crashedSessionHandler endCurrentSessionAsCrashedWhenCrashOrOOM];
 
         [installation sendAllReports];
     };
@@ -161,7 +167,7 @@ SentryCrashIntegration ()
             [deviceData setValue:systemInfo[@"cpuArchitecture"] forKey:@"arch"];
             [deviceData setValue:systemInfo[@"machine"] forKey:@"model"];
             [deviceData setValue:systemInfo[@"model"] forKey:@"model_id"];
-            [deviceData setValue:systemInfo[@"freeMemory"] forKey:@"free_memory"];
+            [deviceData setValue:systemInfo[@"freeMemory"] forKey:SentryDeviceContextFreeMemoryKey];
             [deviceData setValue:systemInfo[@"usableMemory"] forKey:@"usable_memory"];
             [deviceData setValue:systemInfo[@"memorySize"] forKey:@"memory_size"];
             [deviceData setValue:systemInfo[@"storageSize"] forKey:@"storage_size"];